Code/yelp data preprocessing/requirements.txt
textblob
mysql-connector
nltk